refactor(failover): split the repository file by concern
failover.go held the FailoverGroup repository's CRUD alongside the
auto-group sync machinery and the custom-group revalidation. Same
package, pure moves with their doc comments:

- sync.go: SyncResult and its info structs, mergePriorityOrder,
  normalizeBaseModel, deleteUndersizedAutoGroup, upsertAutoGroup,
  diffGroupMembership, SyncAllModels, SyncForModel, PruneModelUUID,
  deleteAutoGroup
- revalidate.go: pruneStaleEntries, pruneMembership, routableMemberIDs,
  revalidateCustomGroups, RevalidateCustomGroups, RevalidateCustomGroupsIn

failover.go keeps the type, the constructor and the CRUD (1091 -> 381
lines) and leaves the size allowlist. The tests were already split
along the same lines (crud_test, repository_syncall_test,
revalidate_test) and are unchanged.
